refactor(team-store): compose collaboration data slice
