+++ test/renderer/store/teamSlice.test.ts
@@ -537,6 +537,34 @@ describe('teamSlice actions', () => {
     vi.useRealTimers();
   });
 
+  it('preserves the collaboration data-plane state and action surface', () => {
+    const state = createSliceStore().getState();
+
+    expect(state).toMatchObject({
+      teams: [],
+      selectedTeamData: null,
+      teamMessagesByName: {},
+      sendingMessage: false,
+      deletedTasks: [],
+      addingComment: false,
+    });
+    expect([
+      state.fetchTeams,
+      state.selectTeam,
+      state.refreshTeamMessagesHead,
+      state.sendTeamMessage,
+      state.createTeamTask,
+      state.addTaskComment,
+    ]).toEqual([
+      expect.any(Function),
+      expect.any(Function),
+      expect.any(Function),
+      expect.any(Function),
+      expect.any(Function),
+      expect.any(Function),
+    ]);
+  });
+
   it('keeps an explicit project intent when opening Teams and clears it for generic navigation', () => {
     const store = createSliceStore();
     store.setState({ selectedProjectId: 'worktree-alpha' });
